Improved handling of class tokens
Proposed method of ReplicaService.ReturnClassToken as a way of checking whether a ClassToken already exists
Proposed clearing class token names when replicas are destroyed, because class tokens are supposed to be unique to the replica and the token is not used when the replica is destroyed

(Context: use case that inspired this was a project where replicas are assigned to the player when they join ex:  ClassToken = ReplicaService.NewClassToken(Replica_[playername])" and when they leave the replica is destroyed. If the player rejoins the game, an error occurs, because the class token already exists.  Two proposed methods, one to check if a class token exists by retrieving it and the other to remove class tokens from internal table when replicas are destroyed)